script to create links for clean pipeline input folder

+++ link_required_files.py
@@ -11,7 +11,7 @@ endings_to_copy = [
 
 input_patient_folders = [f.path for f in os.scandir(input_folder) if f.is_dir()]
 
-for patient_input_folder in input_patient_folders[:3]:
+for patient_input_folder in input_patient_folders:
     # setup input and output folders
     patient_id = os.path.basename(patient_input_folder)
     output_patient_folder = os.path.join(output_folder, patient_id)
@@ -24,7 +24,9 @@ for patient_input_folder in input_patient_folders[:3]:
         file_name = os.path.basename(input_file)
         for ending in endings_to_copy:
             if file_name.endswith(ending):
+                
                 # then create link to the file
                 output_file = os.path.join(output_patient_folder, file_name)
-                input_full_path = os.path.abspath(input_file)
-                os.symlink(input_full_path, output_file)
+                if not os.path.exists(output_file):
+                    input_full_path = os.path.abspath(input_file)
+                    os.symlink(input_full_path, output_file)
